Finish removing containers directory; add Redirect to createTrip action
